requirements.txt
pyrogram==2.0.106
tgcrypto
aiohttp
aiofiles
python-dotenv
